clone churchill into shared vagrant folder
When cloning the churchill repo, save it to the vagrant directory.
This directory is shared with the host. This will allows modification
of the files using any host program.

A symlink is created from /vagrant/churchill to /home/node-user
This allows the churchill-node service running as node-user
to access these files.